Unified Angle Workflow with target labels and cell image resizing
